extract/refactor markdown components
