+++ components/message/Message.tsx
@@ -6,8 +6,9 @@ import { useRouter } from 'next/navigation'
 
 import { DotsThreeX } from '@/components/icons/DotsThreeX'
 import { useMarbleProperties } from '@/components/marble-avatar/Marble'
+import { Markdown } from '@/components/markdown/Markdown'
+import { Pre } from '@/components/markdown/Pre'
 import { Gallery } from '@/components/message/Gallery'
-import { Markdown } from '@/components/message/Markdown'
 import { MessageEditor } from '@/components/message/MessageEditor'
 import { IconButton } from '@/components/ui/Button'
 import { ErrorCallout } from '@/components/ui/Callouts'
@@ -15,7 +16,6 @@ import { Link } from '@/components/ui/Link'
 import { LoadingSpinner } from '@/components/ui/LoadingSpinner'
 import { SkeletonPulse } from '@/components/ui/Skeleton'
 import { AdminOnlyUi } from '@/components/util/AdminOnlyUi'
-import { Pre } from '@/components/util/Pre'
 import { getMessageName, getMessageText } from '@/convex/shared/helpers'
 import { useDeleteMessage } from '@/lib/api'
 import { cn } from '@/lib/utils'
